# WebAPI/WebAPI.Common/Dto/EndangeredAnimalDto.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace WebAPI.Common.Dto
{
    public class EndangeredAnimalDto
    {
        public string Id    { get; set; }
        public string Name { get; set; }
        
    }
}